feat: Expand equipment catalog — 9 new categories, 150 brands, 2,370 models
Add comprehensive home equipment reference data covering:
- HVAC (358 models): Carrier, Trane, Lennox, Goodman, Rheem, Mitsubishi, etc.
- Water Heaters (286 models): A.O. Smith, Rheem, Rinnai, Bradford White, Navien, etc.
- Laundry (211 models): Whirlpool, LG, Samsung, Speed Queen, Miele, Bosch, etc.
- Generators & Power Stations (196 models): Generac, Honda, EcoFlow, Jackery, etc.
- Sump Pumps (118 models): Zoeller, Liberty, Wayne, Basement Watchdog, etc.
- Well Water & Treatment (112 models): Goulds, Grundfos, Culligan, Kinetico, etc.
- Bathroom Fixtures (242 models): Kohler, Moen, Delta, TOTO, Hansgrohe, etc.
- Irrigation (102 models): Rain Bird, Hunter, Rachio, Toro, Orbit, Watts, etc.
- Pool Systems (170 models): Pentair, Hayward, Jandy, Maytronics, etc.

Also includes:
- 107 equipment categories with sub-categories
- 150 manufacturers with support URLs and phone numbers
- 72 common issues with symptoms, fixes, and cost estimates
- 144 service schedule entries with parts lists and frequencies
- Migration and all 14 kitchen seed files applied to Supabase

+++ catalog-data/10_catalog_mainstream.sql
@@ -29,7 +29,7 @@ WHERE m.slug = 'whirlpool' AND c.slug = v.cat_slug;
 -- ======================== MAYTAG ========================
 
 INSERT INTO equipment_catalog (manufacturer_id, category_id, model_number, model_name, series, fuel_type, installation_type, width_inches, capacity_value, capacity_unit, is_current_model, expected_lifespan_years, wifi_enabled, energy_star, msrp_usd, key_features, specs, product_url)
-SELECT m.id, c.id, v.model_number, v.model_name, NULL, v.fuel, v.install, v.width, v.cap, v.cap_unit, true, v.lifespan, false, true, v.msrp, v.features, v.specs::jsonb, 'https://www.maytag.com'
+SELECT m.id, c.id, v.model_number, v.model_name, NULL, v.fuel, v.install, v.width, v.cap, v.cap_unit, true, v.lifespan, false, true, v.msrp::numeric, v.features, v.specs::jsonb, 'https://www.maytag.com'
 FROM equipment_manufacturers m, equipment_categories c,
 (VALUES
   -- French Door Refrigerators
